requirements.txt
arc-alkali-rydberg-calculator
numpy
matplotlib
PyQt6